requirements.txt
rich>=13.0.0
python-nmap>=0.7.1
reportlab>=4.0.4
requests>=2.31.0
beautifulsoup4>=4.12.2
colorama>=0.4.6
